# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Travel planner API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trips:
    get:
      operationId: listTrips
      tags: [trips]
      summary: List trips for the current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trip"
    post:
      operationId: createTrip
      tags: [trips]
      summary: Create a new trip
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
  /trips/{tripId}:
    get:
      operationId: getTrip
      tags: [trips]
      summary: Get trip detail
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "404":
          description: Not found
    patch:
      operationId: updateTrip
      tags: [trips]
      summary: Update trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
    delete:
      operationId: deleteTrip
      tags: [trips]
      summary: Delete trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No content
  /trips/{tripId}/summary:
    get:
      operationId: getTripSummary
      tags: [trips]
      summary: Get trip summary with budget vs expense totals
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripSummary"
...
  /trips/{tripId}/transportation:
    get:
      operationId: listTransportation
      tags: [transportation]
      summary: List transportation records for a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transportation"
    post:
      operationId: createTransportation
      tags: [transportation]
      summary: Add transportation to a trip
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransportationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transportation"
  /trips/{tripId}/transportation/{transportationId}:
    get:
      operationId: getTransportation
      tags: [transportation]
      summary: Get transportation detail
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: transportationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transportation"
    patch:
      operationId: updateTransportation
      tags: [transportation]
      summary: Update transportation
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: transportationId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransportationUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transportation"
    delete:
      operationId: deleteTransportation
      tags: [transportation]
      summary: Delete transportation
      parameters:
        - name: tripId
          in: path
          required: true
          schema:
            type: integer
        - name: transportationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: No content

  # Accommodation
...
components:
  schemas:
...
    Trip:
      type: object
      required: [tripId, userId, title, tripDate, createdAt, updatedAt]
      properties:
        tripId:
          type: integer
        userId:
          type: string
        title:
          type: string
        tripDate:
          type: string
          format: date
        memo:
          type: ["string", "null"]
        companions:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TripInput:
      type: object
      required: [title, tripDate]
      properties:
        title:
          type: string
          maxLength: 200
        tripDate:
          type: string
          format: date
        memo:
          type: string
        companions:
          type: string
    TripUpdate:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
        tripDate:
          type: string
          format: date
        memo:
          type: ["string", "null"]
        companions:
          type: ["string", "null"]
    TripSummary:
      type: object
      required: [tripId, title, tripDate, totalBudget, totalExpense, remaining, categoryBreakdown]
      properties:
        tripId:
          type: integer
        title:
          type: string
        tripDate:
          type: string
          format: date
        totalBudget:
          type: number
        totalExpense:
          type: number
        remaining:
          type: number
        categoryBreakdown:
          type: array
          items:
            $ref: "#/components/schemas/CategoryBudgetExpense"
    CategoryBudgetExpense:
      type: object
      required: [categoryId, categoryName, budget, expense]
      properties:
        categoryId:
          type: integer
        categoryName:
          type: string
        budget:
          type: number
        expense:
          type: number
...
    Transportation:
      type: object
      required: [transportationId, tripId, transportationType, departurePlace, arrivalPlace, departureTime, arrivalTime, fare]
      properties:
        transportationId:
          type: integer
        tripId:
          type: integer
        transportationType:
          type: string
        departurePlace:
          type: string
        arrivalPlace:
          type: string
        departureTime:
          type: string
          format: date-time
        arrivalTime:
          type: string
          format: date-time
        fare:
          type: number
    TransportationInput:
      type: object
      required: [transportationType, departurePlace, arrivalPlace, departureTime, arrivalTime, fare]
      properties:
        transportationType:
          type: string
          maxLength: 50
        departurePlace:
          type: string
          maxLength: 100
        arrivalPlace:
          type: string
          maxLength: 100
        departureTime:
          type: string
          format: date-time
        arrivalTime:
          type: string
          format: date-time
        fare:
          type: number
    TransportationUpdate:
      type: object
      properties:
        transportationType:
          type: string
          maxLength: 50
        departurePlace:
          type: string
          maxLength: 100
        arrivalPlace:
          type: string
          maxLength: 100
        departureTime:
          type: string
          format: date-time
        arrivalTime:
          type: string
          format: date-time
        fare:
          type: number
...