# lib/api-zod/src/generated/types/transportation.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Travel planner API
 * OpenAPI spec version: 0.1.0
 */

export interface Transportation {
  transportationId: number;
  tripId: number;
  transportationType: string;
  departurePlace: string;
  arrivalPlace: string;
  departureTime: Date;
  arrivalTime: Date;
  fare: number;
}