# lib/api-zod/src/generated/types/transportationInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Travel planner API
 * OpenAPI spec version: 0.1.0
 */

export interface TransportationInput {
  /** @maxLength 50 */
  transportationType: string;
  /** @maxLength 100 */
  departurePlace: string;
  /** @maxLength 100 */
  arrivalPlace: string;
  departureTime: Date;
  arrivalTime: Date;
  fare: number;
}