# lib/api-zod/src/generated/types/tripUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Travel planner API
 * OpenAPI spec version: 0.1.0
 */

export interface TripUpdate {
  /** @maxLength 200 */
  title?: string;
  tripDate?: Date;
  /** @nullable */
  memo?: string | null;
  /** @nullable */
  companions?: string | null;
}